laptop/requirements.txt
youtube_dl
python-socketio[client]